requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.28.0
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.45.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap3==21.1
django-bootstrap4==21.1
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-cors-middleware==1.5.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-extensions==3.1.5
django-htmlmin==0.11.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.9
docopt==0.6.2
docutils==0.18
gunicorn==20.1.0
hijri-converter==2.2.2
html5lib==1.1
idna==3.3
jdatetime==3.6.4
Js2Py==0.71
oauthlib==3.1.1
packaging==21.2
Pillow==8.4.0
pipwin==0.5.1
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycparser==2.21
Pygments==2.10.0
pyjsparser==2.7.1
PyJWT==2.3.0
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3
sqlparse==0.4.2
stripe==2.63.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
webencodings==0.5.1
whitenoise==5.3.0
